Adds EngineRouteService
