+++ addon/-private/route-ext.js
@@ -1,68 +1,17 @@
-import Ember from 'ember';
 import Route from '@ember/routing/route';
 import { getOwner } from '@ember/application';
-import { assert } from '@ember/debug';
-import { computed } from '@ember/object';
-import { reads } from '@ember/object/computed';
 
-// @HACK: since `getEngineParent` is not exported
-function getEngineParent(engine) {
-  const symbolPrefix = `__ENGINE_PARENT${Ember.GUID_KEY}`;
-  const symbol = Object.keys(engine).find(k => k.startsWith(symbolPrefix));
-  if (!symbol) {
-    return null;
-  }
-  return engine[symbol];
+/*
+  Creates an aliased form of a method that properly resolves external routes.
+*/
+function routerAlias(methodName) {
+  return function _routerAliasMethod(routeName, ...args) {
+    const router = getOwner(this).lookup('service:router');
+    return router[methodName](routeName, ...args);
+  };
 }
 
 Route.reopen({
-  engine: computed(function () {
-    return getOwner(this);
-  }),
-
-  externalRoutes: reads('engine._externalRoutes'),
-
-  mountPoint: reads('engine.mountPoint'),
-
-  rootApplication: computed(function () {
-    let parent = getEngineParent(getOwner(this));
-    while (getEngineParent(parent)) {
-      parent = getEngineParent(parent);
-    }
-    return parent;
-  }),
-
-  externalRouter: computed(function () {
-    return this.rootApplication.lookup('service:router');
-  }),
-
-  getInternalRouteName(internalRouteName) {
-    // https://github.com/ember-engines/ember-engines/blob/ec4d1ae7a413a7e5d9e57a4e3b2e0f0d19a0afcd/addon/components/link-to-component.js#L52-L57
-    if (internalRouteName === 'application') {
-      return this.mountPoint;
-    }
-    return `${this.mountPoint}.${internalRouteName}`;
-  },
-
-  getExternalRouteName(externalRouteName) {
-    assert(
-      `External route '${externalRouteName}' is unknown.`,
-      externalRouteName in this.externalRoutes
-    );
-    return this.externalRoutes[externalRouteName];
-  },
-
-  transitionToExternal(routeName, ...args) {
-    return this.externalRouter.transitionTo(
-      this.getExternalRouteName(routeName),
-      ...args
-    );
-  },
-
-  replaceWithExternal(routeName, ...args) {
-    return this.externalRouter.replaceWith(
-      this.getExternalRouteName(routeName),
-      ...args
-    );
-  }
+  transitionToExternal: routerAlias('transitionToExternal'),
+  replaceWithExternal: routerAlias('replaceWithExternal'),
 });
